# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Single-user local Codex-like Agent API built with Go and CloudWeGo Eino.\nExternal effects always flow through ToolRouter, EffectInference, PolicyEngine, ApprovalCenter, and Executor.",
        "title": "Local Agent API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/skills/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Upload local skill path",
                "parameters": [
                    {
                        "description": "Local skill upload request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_skills.UploadInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/upload-zip": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Upload skill zip package",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Skill zip archive",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Overwrite the same installed version",
                        "name": "force",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_skills.UploadZipResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Get skill detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SkillDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Remove a registered skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_skills.RemovalResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/disable": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Disable skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/enable": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Enable skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/manifest": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Get skill manifest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SkillManifestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/package": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Get skill package metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.SkillPackageInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/run": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Run skill through ToolRouter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Skill run request",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SkillRunRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ToolRouteResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ToolRouteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/test": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Validate a registered skill request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Validation request",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SkillRunRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_skills.SkillValidateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/skills/{id}/validate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Skills"
                ],
                "summary": "Validate a registered skill",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Skill ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Validation request",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.SkillRunRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_skills.SkillValidateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.SkillDetailResponse": {
            "type": "object",
            "properties": {
                "manifest": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.Manifest"
                },
                "package": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillPackageInfo"
                },
                "skill": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                }
            }
        },
...
        "internal_api_handlers.SkillManifestResponse": {
            "type": "object",
            "properties": {
                "manifest": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.Manifest"
                },
                "skill": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                }
            }
        },
        "internal_api_handlers.SkillRunRequestBody": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "internal_api_handlers.ToolRouteResponse": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                }
            }
        },
...
        "local-agent_internal_core.PolicyDecision": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "approval_payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.SkillPackageInfo": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string"
                },
                "installed_at": {
                    "type": "string"
                },
                "package_path": {
                    "type": "string"
                },
                "skill_id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.SkillRegistration": {
            "type": "object",
            "properties": {
                "approval_default": {
                    "type": "string"
                },
                "archive_path": {
                    "type": "string"
                },
                "checksum": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "installed_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "runtime_type": {
                    "type": "string"
                },
                "sandbox_profile": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_tools_skills.ApprovalConfig": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_tools_skills.Manifest": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.ApprovalConfig"
                },
                "description": {
                    "type": "string"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "input_schema": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                },
                "output_schema": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "permissions": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.PermissionsConfig"
                },
                "runtime": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.Runtime"
                },
                "sandbox": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.SandboxConfig"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_tools_skills.RemovalResult": {
            "type": "object",
            "properties": {
                "package": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillPackageInfo"
                },
                "package_deleted": {
                    "type": "boolean"
                },
                "removed": {
                    "type": "boolean"
                },
                "skill_id": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_tools_skills.SkillValidateResponse": {
            "type": "object",
            "properties": {
                "package": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillPackageInfo"
                },
                "skill": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                },
                "status": {
                    "type": "string"
                },
                "validation": {
                    "$ref": "#/definitions/local-agent_internal_tools_skills.ValidationResult"
                }
            }
        },
...
        "local-agent_internal_tools_skills.UploadInput": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_tools_skills.UploadZipResponse": {
            "type": "object",
            "properties": {
                "package": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillPackageInfo"
                },
                "skill": {
                    "$ref": "#/definitions/local-agent_internal_core.SkillRegistration"
                }
            }
        },
...
    }
}